feat: integrate tsup bundler and replace HTTP example with Next.js app
This commit introduces the tsup bundler, enabling the library to be
compatible with both ESM and CommonJS module systems. Additionally, it
replaces the basic HTTP application example with a Next.js application. This
change enhances the development experience by leveraging hot reloading when
the library is linked to the application.