feat: implement widget creation and improve code structure (#1)
* changes to the codebase based alot of things to make it production ready

* refactor: remove redundant changelog sections

This commit removes the redundant changelog sections to ensure they
align with the https://keepachangelog.com/en/1.1.0/ spec.

* refactor: apply prettier formatting to markdown files

* chore: apply prettier on codebase

* refactor: small docstring changes.

* fixed naming issues, removed unused import and bindings.
added docstring to the contextMenu

* rebasing to changes from visual-nodes branch

* minor

* feat: integrate tsup bundler and replace HTTP example with Next.js app

This commit introduces the tsup bundler, enabling the library to be
compatible with both ESM and CommonJS module systems. Additionally, it
replaces the basic HTTP application example with a Next.js application. This
change enhances the development experience by leveraging hot reloading when
the library is linked to the application.

* removed the older example and updated the readme to reflect the changes. Also modified part of the global.css so that graph loads full screen for every resolution.

* update to the documenetation to reflect the changes to the demo file and directory structure.

* fixed formatting using prettier

* adding another option for demo using vite for hot reloading.

* updates with better formating and removing duplication of documentation information.

* minor change with updates to TODO and dependencies check and updates tooling.

* removing the old demo and finalizing the changes with Vite

* update based on the modular apprach for canvaswrapper and removed the contextmenu patch as it doesnt seem to make any significant effect.

* formatting and updates to the documentation to reflect the changes

* removed unused method in nodeGenerator

* refactor: apply some small syntax improvements

This commit applies some small syntax improvements to the docs and
deployment configuration.

* refactor: switch vite demo to typescript

This commit switches the vite demo app to use typescript.

* fix: ensure that widgets are clickable

This commite ensures that the `pointerevents_method` global is set
before the LGraphCanvas is created. This needs to be done since it is
used in the constructor and can lead to strange behavoirs if set after
the canvas initialisation.

* refactor: improve code structure (#7)

This commit improves the code structure.

* refactor(demo): simplify example code

This commit simplifies the example code.

---------

Co-authored-by: Jason <83615043+JJassonn69@users.noreply.github.com>